Testing If You Can Have Multiple Projects In One Repo
I want to test if you are able to have multiple projects saved to the same repo.  At my current understanding you have one project per repo.  You can roll back to a previous iteration if necessary but you can't do this without deleting the work that succeded this iteration.  If this works I would have mutiple versions of the same project saved in the repo.